[BUG] Unnecessary "utbot_abs_error" constant added to the test #405 (#431)
* [BUG] Unnecessary "utbot_abs_error" constant added to the test #405

- use `EXPECT_FLOAT_EQ` and `EXPECT_DOUBLE_EQ` instead of `EXPECT_NEAR`
- remove insertion obsolete `utbot_abs_error` to the code
- support comparison `NAN` and `INFINITY` consts
- make inclusion of `main.c` wrapper with relative path
- remove obsolete `ABS_ERROR`
- printer optimization

+++ server/test/suites/syntax/floats_special.c
@@ -9,7 +9,7 @@ int is_nanf(float x) {
     }
 }
 
-int is_nan(double x) {
+int is_nan(long double x) {
     if (x != x) {
         return 1;
     } else {
