+++ server/test/suites/syntax/floats_special.h
@@ -2,7 +2,7 @@
 #define UNITTESTBOT_FLOATS_SPECIAL_H
 
 int is_nanf(float x);
-int is_nan(double x);
+int is_nan(long double x);
 int is_inf(float x);
 
 #endif //UNITTESTBOT_FLOATS_SPECIAL_H
